Optimized build_planar_surfaces by using numba and avoided copying arrays as PlanarSurface attributes. The effect is a drastic improvement in calculations with few sites. For instance for Guillaume's disaggregation in oq-risk-tests on my workstation:
```
# master
| calc_16463, maxmem=2.7 GB  | time_sec  | memory_mb | counts  |
|----------------------------+-----------+-----------+---------|
| total classical            | 3_085     | 47.7      | 95      |
| make_contexts              | 2_998     | 43.1      | 30_645  |
| iter_ruptures              | 2_932     | 0.0       | 30_645  |
| ClassicalCalculator.run    | 517.5     | 234.3     | 1       |

# this branch
| calc_16462, maxmem=2.8 GB  | time_sec  | memory_mb | counts  |
|----------------------------+-----------+-----------+---------|
| total classical            | 971.1     | 32.5      | 95      |
| make_contexts              | 886.0     | 28.1      | 30_645  |
| iter_ruptures              | 816.1     | 0.0       | 30_645  |
| ClassicalCalculator.run    | 176.2     | 226.9     | 1       |
```
This is a 3.6x speedup in `iter_ruptures`, but we can do even better.